Re-organization and renaming
Bit of a refactor to put related stuff in their own python modules
