+++ locks.py
@@ -1,0 +1,4 @@
+from threading import Lock
+
+rcon:Lock = Lock()
+testers:Lock = Lock()
