+++ threads.py
@@ -1,0 +1,3 @@
+from testing import RCONThread
+
+rcon:RCONThread = RCONThread()
